去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@28458 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sdbDpsDump.cpp
@@ -1028,7 +1028,7 @@ INT32 _dpsDumper::_analysisMeta()
 
       {
          fs::path fileDir( dirPath ) ;
-         for ( index = 0; index < fileCount; ++index )
+         for ( index = 0; index < (UINT32)fileCount; ++index )
          {
             const CHAR *filepath = fileDir.string().c_str() ;
             CHAR filename[ OSS_MAX_PATHSIZE * 2 ] = { 0 } ;
